refactor: refactor files, implement typescript
