--- src/components/closeFriend/closefriend.css
@@ -1,19 +1,0 @@
-/*.sideBarFriendList {
-    padding: 0;
-    margin: 0;
-    list-style: none;
-}
-
-.sideBarFriend {
-    display: flex;
-    align-items: center;
-    margin-bottom: 15px;
-} 
-
-.sideBarFriendImg {
-    width: 2.25rem;
-    height: 2.25rem;
-    border-radius: 50%;
-    object-fit: cover;
-    margin-right: 2.5rem;
-} */
