--- src/components/feed/feed.css
@@ -1,7 +1,0 @@
-/*.feed {
-    flex: 5.5;
-}
-
-.feedWrapper {
-    padding: 1.25rem;
-} */
