--- src/components/onlineFriend/onlinefriend.css
@@ -1,33 +1,0 @@
-/* .rightBarFriend {
-    display: flex;
-    align-items: center;
-    margin-bottom: 1rem;
-    cursor: pointer;
-} 
-
-.rightBarProfileImgContainer {
-    margin-right: 10px;
-    position: relative;
-}
-
-.rightBarProfileImg {
-    width: 2.25ren;
-    height: 2.25rem;
-    border-radius: 50%;
-    object-fit: cover;
-} 
-
-.rightBarOnline {
-    width: 0.625rem;
-    height: 0.625rem;
-    border-radius: 50%;
-    background-color: limegreen;
-    position: absolute;
-    top: 0px;
-    right: 0px;
-    border: 2px solid white;
-} 
-
-.rightBarUserName {
-    font-weight: 500;
-} */
