--- src/components/post/post.css
@@ -1,82 +1,0 @@
-/*.post {
-    width: 100%;
-    border-radius: 0.75rem;
-    -webkit-box-shadow: 0px 0px 16px -8px rgba(0,0,0,0.68);
-    box-shadow: 0px 0px 16px -8px rgba(0,0,0,0.68);
-    margin: 2rem 0px;
-} 
-
-.postWrapper {
-    padding: 0.875rem;
-} 
-
-.postTop {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-} 
-
-.postTopLeft {
-    display: flex;
-    align-content: center;
-} 
-
-.postProfileImg {
-    width: 2.25rem;
-    height: 2.25rem;
-    border-radius: 50%;
-    object-fit: cover;
-} 
-
-.postUserName {
-    font-size: 1rem;
-    font-weight: 500;
-    margin: 0px 10px;
-    display: flex;
-    align-items: center;
-} 
-
-.postDate {
-    font-size: 12px;
-    color: rgb(113 113 122);
-    display: flex;
-    align-items: center;
-} 
-
-.postCenter {
-    margin: 1.25rem 0px;
-} 
-
-.postImg {
-    margin-top: 1.25rem;
-    width: 100%;
-    max-height: 112rem;
-    object-fit: contain;
-} 
-
-.postBottom {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-} 
-
-.postBottomLeft {
-    display: flex;
-    align-items: center;
-} 
-
-.likeIcon {
-    width: 1.5rem;
-    height: 1.5rem;
-    margin-right: 0.375rem;
-    cursor: pointer;
-} 
-
-.postLikeCounter {
-    font-size: 15px;
-} 
-
-.postCommentText {
-    cursor: pointer;
-    font-size: 15px;
-} */
