--- src/components/rightbar/following/following.css
@@ -1,13 +1,0 @@
-/* .rightBarFollowing {
-    display: flex;
-    flex-direction: column;
-    cursor: pointer;
-    justify-content: center;
-    align-items: center;
-}
-
-.rightBarFollowingImg {
-    width: 100%;
-    height: 100%;
-    object-fit: contain;
-} */
