--- src/components/rightbar/rightbar.css
@@ -1,48 +1,0 @@
-/* .birthDayContainer {
-    display: flex;
-    align-items: center;
-} 
-
-.birthDayImg {
-    width: 2.5rem;
-    height: 2.5rem;
-    margin-right: 0.625rem;
-} 
-
-.birthDayText {
-    font-weight: 300;
-    font-size: 15px;
-}
-
-.rightBarAd {
-    width: 100%;
-    border-radius: 0.75rem;
-    margin: 2rem 0;
-}
-
-.rightBarTitle {
-    font-size: 1.125rem;
-    font-weight: 500;
-    margin-bottom: 1.25rem;
-}
-
-.rightBarInfo {
-    margin-bottom: 2rem;
-}
-
-.rightBar {
-    flex: 3.5;
-    top: 3.5rem;
-}
-
-.rightBarFollowingContainer {
-    display: grid;
-    gap: 1.5rem;
-    grid-template-rows: 8rem;
-    grid-template-columns: 8rem 8rem 8rem;;
-}
-
-.rightBarWrapper {
-    padding: 20px 20px 0px 0px;
-}
-*/
