--- src/components/rightbar/userInfo/userinfo.css
@@ -1,13 +1,0 @@
-/*.rightBarInfoItem {
-    margin-bottom: 0.625rem;
-} 
-
-.rightBarInfoKey {
-    font-weight: 500;
-    margin-right: 1rem;
-    color: #555;
-} 
-
-.rightBarInfoValue {
-    font-weight: 300;
-} */
