--- src/components/share/share.css
@@ -1,78 +1,0 @@
-/*.share {
-    width: 100%;
-    height: 17vh;
-    border-radius: 0.75rem;
-    -webkit-box-shadow: 0px 0px 16px -8px rgba(0,0,0,0.68);
-    box-shadow: 0px 0px 16px -8px rgba(0,0,0,0.68);
-}
-
-.shareWrapper {
-    padding: 0.625rem;
-}  
-
-.shareTop {
-    display: flex;
-    align-items: center;
-} 
-
-.shareProfileImg {
-    width: 3.5rem;
-    height: 3.5rem;
-    border-radius: 50%;
-    object-fit: cover;
-    margin-right: 0.625rem;
-} 
-
-.shareInput {
-    width: 100%;
-    resize: none;
-}
-
-.shareInput:focus {
-    outline: none;
-}
-
-.shareHr {
-    margin: 1.25rem;
-}
-
-.shareBottom {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-}
-
-.shareOptions {
-    margin-left: 1.25rem;
-    display: flex;
-}
-
-.shareOption {
-    display: flex;
-    align-items: center;
-    margin-right: 1rem;
-    justify-content: space-between;
-    cursor: pointer;
-}
-
-.shareIcon {
-    font-size: 1.125rem !important;
-    margin-right: 0.25rem;
-}
-
-.shareOptionText {
-    font-style: 14px;
-    font-weight: 500;
-}
-
-.shareButton {
-    border: none;
-    padding: 0.5rem;
-    border-radius: 0.375rem;
-    background-color: rgb(34 197 94);
-    font-weight: 500;
-    margin-right: 1.25rem;
-    cursor: pointer;
-    color: white;
-}
-*/
