--- src/components/sidebar/sidebar.css
@@ -1,44 +1,0 @@
-/*.sideBar {
-    flex: 3;
-    height: calc(100vh - 3.5rem);
-    overflow-y: scroll;
-    position: sticky;
-    top: 3.5rem;
-}
-
-.sideBarWrapper {
-    padding: 1.25rem;
-}
-
-.sideBarList {
-    padding: 0;
-    margin: 0;
-    list-style: none;
-}
-
-.sideBarListItem {
-    display: flex;
-    align-items: center;
-    margin-bottom: 1.25rem;
-}
-
-.sideBarIcon {
-    margin-right: 1.125rem;
-}
-
-.sideBarButton {
-    width: 9rem;
-    border: none;
-    padding: 0.625rem;
-    border-radius: 0.375rem;
-    font-weight: 500;
-}
-
-.sideBarHr {
-    margin: 1.25rem 0;
-}
-
-.sideBarCloseFriendList {
-    margin-bottom: 1rem;
-}
-*/
