--- src/dummyData.ts
@@ -1,299 +1,0 @@
-export const Users = [
-  {
-    id: 1,
-    profilePicture: "assets/person/1.jpeg",
-    userName: "Berk Yaşar",
-    online: true,
-    closeFriend: false,
-    currentUser: true,
-    following: [1,2,3,4,5,6,7,8,9,10,11],
-  },
-  {
-    id: 2,
-    profilePicture: "assets/person/2.jpeg",
-    userName: "Janell Shrum",
-    online: false,
-    closeFriend: true,
-    currentUser: false,
-    following: [1,2,3,4,5,11],
-  },
-  {
-    id: 3,
-    profilePicture: "assets/person/3.jpeg",
-    userName: "Alex Durden",
-    online: true,
-    closeFriend: true,
-  },
-  {
-    id: 4,
-    profilePicture: "assets/person/4.jpeg",
-    userName: "Dora Hawks",
-    online: true,
-    closeFriend: true,
-  },
-  {
-    id: 5,
-    profilePicture: "assets/person/5.jpeg",
-    userName: "Thomas Holden",
-    onine: true,
-    closeFriend: true,
-  },
-  {
-    id: 6,
-    profilePicture: "assets/person/6.jpeg",
-    userName: "Shirley Beauchamp",
-    online: false,
-    closeFriend: true,
-  },
-  {
-    id: 7,
-    profilePicture: "assets/person/7.jpeg",
-    userName: "Travis Bennett",
-    online: false,
-    closeFriend: true,
-  },
-  {
-    id: 8,
-    profilePicture: "assets/person/8.jpeg",
-    userName: "Kristen Thomas",
-    online: true,
-    closeFriend: true,
-  },
-  {
-    id: 9,
-    profilePicture: "assets/person/9.jpeg",
-    userName: "Gary Duty",
-    online: true,
-    closeFriend: true,
-  },
-  {
-    id: 10,
-    profilePicture: "assets/person/lebron-james.jpg",
-    userName: "Lebron James",
-    online: false,
-    closeFriend: true,
-  },
-];
-
-export const Posts = [
-  {
-    id: 1,
-    desc: "Love For All, Hatred For None.",
-    photo: "assets/post/1.jpeg",
-    date: "5 mins ago",
-    userId: 1,
-    like: 54,
-    comment: 9,
-  },
-  {
-    id: 2,
-    photo: "assets/post/2.jpeg",
-    date: "15 mins ago",
-    userId: 2,
-    like: 2,
-    comment: 1,
-  },
-  {
-    id: 3,
-    desc: "Every moment is a fresh beginning.",
-    photo: "assets/post/3.jpeg",
-    date: "1 hour ago",
-    userId: 3,
-    like: 61,
-    comment: 2,
-  },
-  {
-    id: 4,
-    photo: "assets/post/4.jpeg",
-    date: "4 hours ago",
-    userId: 4,
-    like: 7,
-    comment: 3,
-  },
-  {
-    id: 5,
-    photo: "assets/post/5.jpeg",
-    date: "5 hours ago",
-    userId: 5,
-    like: 23,
-    comment: 5,
-  },
-  {
-    id: 6,
-    photo: "assets/post/6.jpeg",
-    date: "1 day ago",
-    userId: 6,
-    like: 44,
-    comment: 6,
-  },
-  {
-    id: 7,
-    desc: "Never regret anything that made you smile.",
-    photo: "assets/post/7.jpeg",
-    date: "2 days ago",
-    userId: 7,
-    like: 52,
-    comment: 3,
-  },
-  {
-    id: 8,
-    photo: "assets/post/8.jpeg",
-    date: "3 days ago",
-    userId: 8,
-    like: 15,
-    comment: 1,
-  },
-  {
-    id: 9,
-    desc: "Change the world by being yourself.",
-    photo: "assets/post/9.jpeg",
-    date: "5 days ago",
-    userId: 9,
-    like: 11,
-    comment: 2,
-  },
-  {
-    id: 10,
-    photo: "assets/post/10.jpeg",
-    date: "1 week ago",
-    userId: 10,
-    like: 104,
-    comment: 12,
-  },
-];
-
-export const UsersProfileDetails = [
-  {
-    id: 1,
-    city: 'New York',
-    from: 'Madrid',
-    relationShip: 'Married',
-    birthDay: '18/08/1995 (28 Years old)',
-    job: 'Frontend Developer',
-    school: 'Istanbul University',
-    profileDescription: 'Love you all, welcome to my page!',
-  },
-  {
-    id: 2,
-    city: 'New York',
-    from: 'Argentina',
-    relationShip: 'Complicated',
-    job: 'Student',
-    school: 'New York University',
-    profileDescription: 'Hello there!',
-  },
-  {
-    id: 3,
-    city: 'New York',
-    from: 'Brazil',
-    relationShip: 'Married',
-    birthDay: '16/10/1995 (28 Years old)',
-    job: 'QA Engineer',
-    profileDescription: 'I dont know why i am here',
-  },
-  {
-    id: 4,
-    city: 'New York',
-    from: 'Turkey',
-    relationShip: 'Single',
-    birthDay: '13/02/2000 (23 Years old)',
-    job: 'Graphical Designer',
-  },
-  {
-    id: 5,
-    city: 'New York',
-    from: 'Netherlands',
-    relationShip: 'Married',
-    birthDay: '08/12/1979 (44 Years old)',
-    job: 'Scrum Master',
-    profileDescription: 'Lots of Love!',
-  },
-  {
-    id: 6,
-    city: 'New York',
-    from: 'Texas',
-  },
-  {
-    id: 7,
-    city: 'New York',
-    from: 'Canada',
-  },
-  {
-    id: 8,
-    job: 'Frontend Developer',
-    profileDescription: 'Born to code!',
-  },
-  {
-    id: 9,
-    city: 'Madrid',
-    from: 'Madrid',
-    relationShip: 'Married',
-    birthDay: '22/11/1995 (28 Years old)',
-  },
-  {
-    id: 10,
-    city: 'Madrid',
-    from: 'Madrid',
-    relationShip: 'Married',
-    birthDay: '03/05/1995 (28 Years old)',
-    profileDescription: 'My album is out soon, stay tuned!',
-  },
-];
-
-export const Companies = [
-  {
-    id: 1,
-    name: 'Apple',
-    logo: '/assets/logo/apple.png',
-  },
-  {
-    id: 2,
-    name: 'Ac/Dc',
-    logo: '/assets/logo/acdc.png',
-  },
-  {
-    id: 3,
-    name: 'Audi',
-    logo: '/assets/logo/audi.png',
-  },
-  {
-    id: 4,
-    name: 'CocaCola',
-    logo: '/assets/logo/cola.png',
-  },
-  {
-    id: 5,
-    name: 'Facebook',
-    logo: '/assets/logo/facebook.png',
-  },
-  {
-    id: 6,
-    name: 'Guns N Roses',
-    logo: '/assets/logo/gnr.png',
-  },
-  {
-    id: 7,
-    name: 'Google',
-    logo: '/assets/logo/google.png',
-  },
-  {
-    id: 8,
-    name: 'Mercedes',
-    logo: '/assets/logo/mercedes.png',
-  },
-  {
-    id: 9,
-    name: 'Porsche',
-    logo: '/assets/logo/porsche.png',
-  },
-  {
-    id: 10,
-    name: 'Tesla',
-    logo: '/assets/logo/tesla.png',
-  },
-  {
-    id: 11,
-    name: 'Twitter',
-    logo: '/assets/logo/twitter.png',
-  },
-];
