+++ src/helper/api/companies.ts
@@ -1,0 +1,60 @@
+import { Company } from "../types/company";
+
+export const companies: Company[] = [
+    {
+      id: 1,
+      name: 'Apple',
+      logo: '/assets/logo/apple.png',
+    },
+    {
+      id: 2,
+      name: 'Ac/Dc',
+      logo: '/assets/logo/acdc.png',
+    },
+    {
+      id: 3,
+      name: 'Audi',
+      logo: '/assets/logo/audi.png',
+    },
+    {
+      id: 4,
+      name: 'CocaCola',
+      logo: '/assets/logo/cola.png',
+    },
+    {
+      id: 5,
+      name: 'Facebook',
+      logo: '/assets/logo/facebook.png',
+    },
+    {
+      id: 6,
+      name: 'Guns N Roses',
+      logo: '/assets/logo/gnr.png',
+    },
+    {
+      id: 7,
+      name: 'Google',
+      logo: '/assets/logo/google.png',
+    },
+    {
+      id: 8,
+      name: 'Mercedes',
+      logo: '/assets/logo/mercedes.png',
+    },
+    {
+      id: 9,
+      name: 'Porsche',
+      logo: '/assets/logo/porsche.png',
+    },
+    {
+      id: 10,
+      name: 'Tesla',
+      logo: '/assets/logo/tesla.png',
+    },
+    {
+      id: 11,
+      name: 'Twitter',
+      logo: '/assets/logo/twitter.png',
+    },
+  ];
+
