+++ src/helper/api/profileDetails.ts
@@ -1,0 +1,79 @@
+import { ProfileDetail } from "../types/profileDetail";
+
+export const usersProfileDetails: ProfileDetail[] = [
+  {
+    id: 1,
+    city: 'New York',
+    from: 'Madrid',
+    relationship: 'Married',
+    birthday: '18/08/1995 (28 Years old)',
+    job: 'Frontend Developer',
+    school: 'Istanbul University',
+    profileDescription: 'Love you all, welcome to my page!',
+  },
+  {
+    id: 2,
+    city: 'New York',
+    from: 'Argentina',
+    relationship: 'Complicated',
+    job: 'Student',
+    school: 'New York University',
+    profileDescription: 'Hello there!',
+  },
+  {
+    id: 3,
+    city: 'New York',
+    from: 'Brazil',
+    relationship: 'Married',
+    birthday: '16/10/1995 (28 Years old)',
+    job: 'QA Engineer',
+    profileDescription: 'I dont know why i am here',
+  },
+  {
+    id: 4,
+    city: 'New York',
+    from: 'Turkey',
+    relationship: 'Single',
+    birthday: '13/02/2000 (23 Years old)',
+    job: 'Graphical Designer',
+  },
+  {
+    id: 5,
+    city: 'New York',
+    from: 'Netherlands',
+    relationship: 'Married',
+    birthday: '08/12/1979 (44 Years old)',
+    job: 'Scrum Master',
+    profileDescription: 'Lots of Love!',
+  },
+  {
+    id: 6,
+    city: 'New York',
+    from: 'Texas',
+  },
+  {
+    id: 7,
+    city: 'New York',
+    from: 'Canada',
+  },
+  {
+    id: 8,
+    job: 'Frontend Developer',
+    profileDescription: 'Born to code!',
+  },
+  {
+    id: 9,
+    city: 'Madrid',
+    from: 'Madrid',
+    relationship: 'Married',
+    birthday: '22/11/1995 (28 Years old)',
+  },
+  {
+    id: 10,
+    city: 'Madrid',
+    from: 'Madrid',
+    relationship: 'Married',
+    birthday: '03/05/1995 (28 Years old)',
+    profileDescription: 'My album is out soon, stay tuned!',
+  },
+];
