+++ src/helper/types/company.ts
@@ -1,0 +1,5 @@
+export interface Company {
+    id: number;
+    name: string;
+    logo: string;
+}
