+++ src/helper/types/post.ts
@@ -1,0 +1,10 @@
+export interface Post {
+    id: number;
+    desc?: string;
+    photo?: string;
+    date?: string;
+    userId?: number;
+    like: number;
+    comment?: number;
+}
+
