+++ src/helper/types/profileDetail.ts
@@ -1,0 +1,11 @@
+export interface ProfileDetail {
+    id: number;
+    city?: string;
+    from?: string;
+    relationship?: string;
+    birthday?: string;
+    info?: string;
+    job?: string;
+    school?: string;
+    profileDescription?: string;
+}
