+++ src/helper/types/user.ts
@@ -1,0 +1,10 @@
+export interface User {
+    id: number;
+    profilePicture?: string;
+    userName?: string;
+    online?: boolean;
+    info?: string;
+    closeFriend?: boolean;
+    currentUser?: boolean;
+    following?: number[];
+}
