--- src/pages/home/home.css
@@ -1,5 +1,0 @@
-/* Without Tailwind */
-/* .homeContainer {
-    display: flex;
-    width: 100%;
-} */
