--- src/pages/profile/profile.css
@@ -1,53 +1,0 @@
-/*
-.profile {
-    display: flex;
-}
-
-.profileRight {
-    flex: 9;
-} 
-
-.profileCover {
-    height: 32vh;
-    position: relative;
-}
-
-.profileCoverImg {
-    width: 100%;
-    height: 80%;
-    object-fit: cover;
-}  
-
-.profileUserImg {
-    width: 36rem;
-    height: 36rem;
-    border-radius: 50%;
-    object-fit: cover;
-    position: absolute;
-    left: 0;
-    right: 0;
-    margin: auto;
-    top: 50%;
-    cursor: pointer;
-    border: 4px solid white;
-} 
-
-.profileInfo {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-
-} 
-
-.profileInfoName {
-    font-size: 24px;
-}
-
-.profileInfoDesc {
-    font-weight: 300;
-}
-
-.profileRightBottom {
-    display: flex;
-} */
