+++ tsconfig.json
@@ -1,5 +1,6 @@
 {
     "compilerOptions": {
+        "jsx": "react",
         "target": "ES2020",
         "lib": [
             "dom",
@@ -18,7 +19,6 @@
         "resolveJsonModule": true,
         "isolatedModules": true,
         "noEmit": true,
-        "jsx": "react-jsx"
     },
     "include": [
         "src"
